+++ NEWS.md
@@ -1,5 +1,9 @@
 # History of user-visible changes
 
+## Next
+
+* New user option `company-etags-ignore-case`.
+
 ## 2014-03-19 (0.7.2)
 
 * Support for Emacs 22 officially dropped.